+++ mod.ts
@@ -34,6 +34,6 @@ export {
 /** types */
 export type { InitializerFunction, ResolverFunction } from "./polymorphic.ts";
 
-export { getNew } from "./strategy/utils.ts";
+export { getNewSerializable } from "./strategy/utils.ts";
 /** types */
 export type { SerializableConstructor } from "./strategy/utils.ts";