+++ strategy/from_json/to_serializable.ts
@@ -2,7 +2,7 @@
 
 import { JSONValue, Serializable } from "../../serializable.ts";
 import { FromJSONStrategy } from "../compose_strategy.ts";
-import { getNew, SerializableConstructor } from "../utils.ts";
+import { getNewSerializable, SerializableConstructor } from "../utils.ts";
 
 /** revive data using `fromJSON` on a subclass type */
 export function toSerializable<T>(
@@ -12,8 +12,8 @@ export function toSerializable<T>(
     value: JSONValue,
   ): Serializable | Serializable[] {
     if (Array.isArray(value)) {
-      return value.map((item) => getNew(type).fromJSON(item));
+      return value.map((item) => getNewSerializable(type).fromJSON(item));
     }
-    return getNew(type).fromJSON(value);
+    return getNewSerializable(type).fromJSON(value);
   };
 }