+++ strategy/utils_test.ts
@@ -1,13 +1,13 @@
 // Copyright 2018-2021 Gamebridge.ai authors. All rights reserved. MIT license.
 
 import { test } from "../test_deps.ts";
-import { getNew } from "./utils.ts";
+import { getNewSerializable } from "./utils.ts";
 import { assertEquals } from "../test_deps.ts";
 import { Serializable } from "../serializable.ts";
 import { SerializeProperty } from "../serialize_property.ts";
 
 test({
-  name: "getNew takes a function that returns a Serializable",
+  name: "getNewSerializable takes a function that returns a Serializable",
   fn() {
     class A extends Serializable {
       @SerializeProperty()
@@ -19,7 +19,7 @@ test({
     }
 
     assertEquals(
-      getNew(
+      getNewSerializable(
         () => new A({ test: "from_constructor" }),
       ).toJSON(),
       `{"test":"from_constructor"}`,
@@ -28,7 +28,7 @@ test({
 });
 
 test({
-  name: "getNew takes a Serializable",
+  name: "getNewSerializable takes a Serializable",
   fn() {
     class A extends Serializable {
       @SerializeProperty()
@@ -36,7 +36,7 @@ test({
     }
 
     assertEquals(
-      getNew(A).toJSON(),
+      getNewSerializable(A).toJSON(),
       `{"test":"not_from_constructor"}`,
     );
   },